requirements.txt
python==3.7
pytorch==1.9.0
bert-score
dataclasses==0.6
datasets==1.0.2
matplotlib==3.1.3
nlp==0.4.0
nltk==3.4.5
numba==0.51.2
numpy==1.18.1
overrides==2.8.0
pandas==0.25.3
scikit-learn==0.21.3
scipy==1.4.1
seaborn==0.10.0
sentencepiece==0.1.85
tqdm==4.43.0
transformers==3.1.0
urllib3>=1.26.5
websocket-client==0.57.0
